@@ -1,11 +1,14 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+
+import { AppComponent } from './app.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 
 import { CoreModule } from './core/core.module';
 import { SharedModule } from './shared/shared.module';
-import { AppComponent } from './app.component';
 import { ConnectModule } from './connect/connect.module';
 import { ConnectComponent } from './connect/connect.component';
 
@@ -14,8 +17,9 @@ import { ConnectComponent } from './connect/connect.component';
     AppComponent
   ],
   imports: [
-    CoreModule,
     BrowserModule,
+    BrowserAnimationsModule,
+    CoreModule,
     SharedModule,
     ConnectModule
   ],
@@ -23,3 +27,5 @@ import { ConnectComponent } from './connect/connect.component';
   bootstrap: [AppComponent]
 })
 export class AppModule { }
+
+
