@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { MaterialModule } from '@angular/material';
 
 @NgModule({
-    imports:        [FormsModule, MaterialModule],
-    exports:        [FormsModule, MaterialModule]
+  imports: [FormsModule],
+  exports: [FormsModule]
 })
-export class SharedModule { }
+export class SharedModule {}
